Refactored Index.js
